platform/msm8916: Rearrange Ramdisk and DTB addresses
We are experimenting data abort handling issues with a ramdisk with
size of 0x1acd5d0 the ramdisk address was set at 0x84000000 with final
address 0x85acd5d0 but crashes when reach 0x85a00000 because the mapping
is for 90 MB (0x84000000 - 0x85a00000).

...
[3230] [3230] data abort, halting
[3230] [3230] r0  0x85a00000 r1  0x924d1010 r2  0x000cd5b0 r3  0x14ad1000
[3230] [3230] r4  0xcf4d1acc r5  0x6bff89f3 r6  0x0155f72e r7  0x8178233a
[3230] [3230] r8  0x009d0800 r9  0x90100000 r10 0x01a63000 r11 0x9259f000
[3230] [3230] r12 0x00000001 usp 0x00000000 ulr 0x00000000 pc  0x8f629c30
[3230] [3230] spsr 0x20000153
...

Since we only have space from 0x80080000 (kernel start) to 0x860000000
(tz zone starts), rearrange to provide more space to the ramdisk.

0x80080000 - 0x82080000, kernel (32MB)
0x82080000 - 0x82280000, dtb (2MB)
0x82280000 - 0x86000000, ramdisk (61.5MB)

Change-Id: I07fd7f982a07c98eb8f783f1aca5991ac6643741
Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>